chore: Further refactoring of the downloaders. chore: Removed unnecessary Rich markup in logger messages.
