+++ src/ollama_downloader/downloader/model_downloader.py
@@ -1,280 +1,388 @@
-import datetime
-import json
+from abc import ABC, abstractmethod
+from enum import Enum
+import hashlib
 import logging
 import os
-from typing import Dict, List, Set, Tuple
+import platform
+import shutil
+import ssl
+import tempfile
+from typing import Set, Tuple
+from urllib.parse import urlparse
 
-from httpx import URL
+import certifi
+import httpx
 
 from environs import env
 
 from ollama_downloader.common import EnvVar
-from ollama_downloader.data.data_models import ImageManifest
-from ollama_downloader.downloader.downloader import Downloader, ModelSource
-import lxml.html
-from ollama import Client as OllamaClient
 
-
-from rich import print as print
 from rich.progress import (
     Progress,
     BarColumn,
     TextColumn,
-    MofNCompleteColumn,
+    DownloadColumn,
+    TransferSpeedColumn,
 )
 
-# Initialize the logger
+from ollama_downloader.data.data_models import AppSettings
+
 logger = logging.getLogger(__name__)
-logger.setLevel(env.str(EnvVar.LOG_LEVEL, default=EnvVar.DEFAULT__LOG_LEVEL).upper())
 
 
-class OllamaModelDownloader(Downloader):
+class ModelSource(Enum):
+    OLLAMA = 1
+    HUGGINGFACE = 2
+
+
+class ModelDownloader(ABC):
+    HF_BASE_URL = "https://hf.co/v2/"
+
     def __init__(self):
-        super().__init__()
-        self.models_tags: Dict[str, list] = {}
-        self.library_models: List[str] = []
+        self.settings: AppSettings = AppSettings.load_or_create_default()
+        if not self.settings:
+            raise RuntimeError("Failed to load or create and save default settings.")
+        self._user_agent: str = f"{env.str(EnvVar.OD_UA_NAME_VER, default=EnvVar.DEFAULT__OD_UA_NAME_VER)} ({platform.platform()} {platform.system()}-{platform.release()} Python-{platform.python_version()})"
+        self._unnecessary_files: Set[str] = set()
+        self._cleanup_running: bool = False
 
-    def _save_models_tags_cache(
-        self,
-    ) -> None:
-        """
-        Save the models tags cache to a file.
-        """
-        if hasattr(self, "models_tags"):
-            with open(self.settings.ollama_library.models_tags_cache, "w") as f:
-                f.write(json.dumps(self.models_tags, indent=2))
-        else:
-            logger.warning("No models tags cache to save.")
-
-    def load_models_tags_cache(self) -> None:
-        """
-        Load the models tags cache from a file.
-        """
-        if os.path.exists(self.settings.ollama_library.models_tags_cache):
-            with open(self.settings.ollama_library.models_tags_cache, "r") as f:
-                self.models_tags = json.loads(f.read())
-            logger.info(
-                f"Loaded models tags cache from {self.settings.ollama_library.models_tags_cache}"
-            )
-            self.library_models = list(self.models_tags.keys())
-        else:
-            logger.warning(
-                f"No models tags cache found at {self.settings.ollama_library.models_tags_cache}"
-            )
-
+    @abstractmethod
     def download_model(self, model_identifier: str) -> bool:
-        model, tag = (
-            model_identifier.split(":")
-            if ":" in model_identifier
-            else (model_identifier, "latest")
-        )
-        print(f"Downloading Ollama library model {model_identifier}")
-        # Validate the response as an ImageManifest but don't enforce strict validation
-        manifest_json = self._fetch_manifest(
-            model_identifier=model_identifier, model_source=ModelSource.OLLAMA
-        )
-        logger.debug(f"Validating manifest for {model}:{tag}")
-        manifest = ImageManifest.model_validate_json(manifest_json, strict=False)
-        logger.info(
-            f"Downloading model configuration [bold cyan]{manifest.config.digest}[/bold cyan]"
-        )
-        # Keep a list of files to be copied but only copy after all downloads have completed successfully.
-        # This is to ensure that we don't copy files that may not be needed if the download fails.
-        # Each tuple in the list contains (source_path, named_digest, computed_digest).
-        files_to_be_copied: List[Tuple[str, str, str]] = []
-        # Download the model configuration BLOB
-        file_model_config, digest_model_config = self._download_model_blob(
-            model_identifier=model_identifier,
-            named_digest=manifest.config.digest,
-            model_source=ModelSource.OLLAMA,
-        )
-        files_to_be_copied.append(
-            (file_model_config, manifest.config.digest, digest_model_config)
-        )
-        for layer in manifest.layers:
-            logger.debug(
-                f"Layer: [bold cyan]{layer.mediaType}[/bold cyan], Size: [bold green]{layer.size}[/bold green] bytes, Digest: [bold yellow]{layer.digest}[/bold yellow]"
-            )
-            logger.info(
-                f"Downloading [bold cyan]{layer.mediaType}[/bold cyan] layer [bold cyan]{layer.digest}[/bold cyan]"
-            )
-            file_layer, digest_layer = self._download_model_blob(
-                model_identifier=model_identifier,
-                named_digest=layer.digest,
-                model_source=ModelSource.OLLAMA,
-            )
-            files_to_be_copied.append((file_layer, layer.digest, digest_layer))
-        # All BLOBs have been downloaded, now copy them to their appropriate destinations.
-        for source, named_digest, computed_digest in files_to_be_copied:
-            copy_status, copy_destination = self._save_blob(
-                source=source,
-                named_digest=named_digest,
-                computed_digest=computed_digest,
-            )
-            if copy_status is False:
-                raise RuntimeError(
-                    f"Failed to copy {named_digest} to {copy_destination}."
-                )
-        # Finally, save the manifest to its appropriate destination
-        self._save_manifest(
-            data=manifest_json,
-            model_identifier=model_identifier,
-            model_source=ModelSource.OLLAMA,
-        )
-        ts_approximate_manifest_save = datetime.datetime.now()
-        # Finally check if it exists in the Ollama
-        # Clear the list of unnecessary files before this if errors henceforth are to be tolerated.
-        if not self.settings.ollama_server.remove_downloaded_on_error:
-            self._unnecessary_files.clear()
-        ollama_client = OllamaClient(
-            host=self.settings.ollama_server.url,
-            # timeout=self.settings.ollama_server.timeout,
-            # TODO: Add API key authentication logic
-        )
-        models_list = ollama_client.list()
-        found_model = None
-        for model_info in models_list.models:
-            if (
-                model_info.model == f"{model}:{tag}"
-                # TODO: Is this timestamp assumption right that the listing is completed within a minute of saving?
-                and abs(
-                    model_info.modified_at.replace(tzinfo=None)
-                    - ts_approximate_manifest_save
-                )
-                < datetime.timedelta(minutes=1)
-            ):
-                found_model = model_info
-                break
-        if found_model:
-            print(
-                f"Model {found_model.model} successfully downloaded and saved on {found_model.modified_at:%B %d %Y at %H:%M:%S}."
-            )
-        else:
-            raise RuntimeError(
-                f"Model {model}:{tag} could not be found in Ollama server after download."
-            )
-        # If we reached here cleanly, remove all unnecessary file names but don't remove actual files.
-        self._unnecessary_files.clear()
-        # Finally return success
-        return True
-
-    def update_models_list(self) -> list:
         """
-        Update the list of models available in the Ollama library.
+        Download a supported model into an available Ollama server.
+
+        Args:
+            model_identifier (str): The model tag to download, e.g., "gpt-oss:latest" for library models.
+            If the tag is omitted, "latest" is assumed. For Hugging Face models, the model identifier is
+            of the format <user>/<repository>:<quantisation>, e.g., unsloth/gemma-3-270m-it-GGUF:Q4_K_M.
 
         Returns:
-            list: A list of model names available in the Ollama library.
+            bool: True if the model was successfully downloaded and verified, False otherwise.
         """
-        with self.get_httpx_client(
-            verify=self.settings.ollama_library.verify_ssl,
-            timeout=self.settings.ollama_library.timeout,
-        ) as client:
-            logger.debug(
-                f"Updating models list from Ollama library {self.settings.ollama_library.library_base_url}"
-            )
-            models_response = client.get(self.settings.ollama_library.library_base_url)
-            models_response.raise_for_status()
-            parsed_models_html = lxml.html.document_fromstring(models_response.text)
-            self.library_models = []
-            library_prefix = "/library/"
-            for _, attribute, link, _ in lxml.html.iterlinks(parsed_models_html):
-                if attribute == "href" and link.startswith(library_prefix):
-                    self.library_models.append(link.replace(library_prefix, ""))
-            logger.debug(
-                f"Found {len(self.library_models)} models in the Ollama library."
-            )
-            return self.library_models
+        pass
 
-    def list_models_tags(
-        self, model: str | None = None, update: bool = False
-    ) -> dict[str, list]:
+    def get_httpx_client(self, verify: bool, timeout: float) -> httpx.Client:
         """
-        Update the tags for each model or a named model in the Ollama library.
+        Obtain an HTTPX client for making requests.
+
+        Args:
+            verify (bool): Whether to verify SSL certificates.
+            timeout (float): The timeout for requests in seconds.
 
         Returns:
-            dict[str, list]: A dictionary where keys are model names and values are lists of tags.
+            httpx.Client: An HTTPX client configured with the specified settings.
         """
-        self.load_models_tags_cache()
-        if not hasattr(self, "library_models") or update:
-            self.update_models_list()
-        _base_url = URL(self.settings.ollama_library.library_base_url)
+        if verify is False:
+            logger.warning(
+                "SSL verification is disabled. This is not recommended for production use."
+            )
+        ctx = ssl.create_default_context(
+            cafile=env.str("SSL_CERT_FILE", default=certifi.where()),
+            capath=env.str("SSL_CERT_DIR", default=None),
+        )
+        client = httpx.Client(
+            verify=verify if (verify is not None and verify is False) else ctx,
+            follow_redirects=True,
+            trust_env=True,
+            http2=True,
+            timeout=timeout,
+            headers={"User-Agent": self._user_agent},
+        )
+        return client
+
+    def _make_manifest_url(
+        self, model_identifier: str, model_source: ModelSource
+    ) -> httpx.URL:
+        """
+        Constructs the manifest URL for a given model identifier.
+
+        Args:
+            model_identifier (str): The model identifier, e.g., "gpt-oss:latest" for an Ollama library model
+            or "unsloth/gemma-3-270m-it-GGUF:Q4_K_M" for a Hugging Face model.
+            model_source (ModelSource): The source of the model (e.g., OLLAMA or HUGGINGFACE).
+
+        Returns:
+            httpx.URL: The constructed manifest URL.
+        """
+        match model_source:
+            case ModelSource.OLLAMA:
+                model, tag = (
+                    model_identifier.split(":")
+                    if ":" in model_identifier
+                    else (model_identifier, "latest")
+                )
+                logger.debug(f"Constructing manifest URL for {model}:{tag}")
+                return httpx.URL(self.settings.ollama_library.registry_base_url).join(
+                    f"{model}/manifests/{tag}"
+                )
+            case ModelSource.HUGGINGFACE:
+                logger.debug(f"Constructing manifest URL for {model_identifier}")
+                return httpx.URL(
+                    f"{ModelDownloader.HF_BASE_URL}{model_identifier.replace(':', '/manifests/')}"
+                )
+            case _:
+                raise ValueError(f"Unsupported model source: {model_source}")
+
+    def _fetch_manifest(self, model_identifier: str, model_source: ModelSource) -> str:
+        """
+        Fetches the manifest JSON content for a given model identifier.
+
+        Args:
+            model_identifier (str): The model identifier, e.g., "gpt-oss:latest" for an Ollama library model
+            or "unsloth/gemma-3-270m-it-GGUF:Q4_K_M" for a Hugging Face model.
+            model_source (ModelSource): The source of the model (e.g., OLLAMA or HUGGINGFACE).
+
+        Returns:
+            str: The fetched manifest JSON content as a string.
+        """
+        url = self._make_manifest_url(model_identifier, model_source)
+        logger.info(f"Downloading manifest from {url}")
         with self.get_httpx_client(
-            verify=self.settings.ollama_library.verify_ssl,
-            timeout=self.settings.ollama_library.timeout,
-        ) as client:
-            if model is not None:
-                if model not in self.library_models:
-                    logger.error(f"Model {model} not found in the library models list.")
-                    return {}
-                else:
-                    if not hasattr(self, "models_tags") or update:
-                        logger.debug(
-                            f"Fetching tags for model {model} from the Ollama library."
-                        )
-                        tags_response = client.get(_base_url.join(f"{model}/tags"))
-                        tags_response.raise_for_status()
-                        logger.debug(f"Parsing tags for model {model}.")
-                        parsed_tags_html = lxml.html.document_fromstring(
-                            tags_response.text
-                        )
-                        library_prefix = "/library/"
-                        named_model_unique_tags: Set[str] = set()
-                        for _, attribute, link, _ in lxml.html.iterlinks(
-                            parsed_tags_html
-                        ):
-                            if attribute == "href" and link.startswith(
-                                f"{library_prefix}{model}:"
-                            ):
-                                if model not in self.models_tags:
-                                    self.models_tags[model] = []
-                                named_model_unique_tags.add(
-                                    link.replace(library_prefix, "")
-                                )
-                        self.models_tags[model] = list(named_model_unique_tags)
-                        logger.debug(f"Updating tags for model {model} in the cache.")
-                        self._save_models_tags_cache()
-                    return {model: self.models_tags.get(model, [])}
-            else:
-                if (
-                    not hasattr(self, "models_tags")
-                    or update
-                    or len(self.models_tags) == 0
-                ):
-                    # A full update has been requested
-                    model_counter = 0
-                    with Progress(
-                        TextColumn(
-                            text_format="[bold blue]{task.description}[/bold blue]"
-                        ),
-                        "[progress.percentage]{task.percentage:>3.0f}%",
-                        BarColumn(bar_width=None),
-                        MofNCompleteColumn(),
-                    ) as progress:
-                        tags_task = progress.add_task(
-                            "Updating models",
-                            total=len(self.library_models),
-                        )
-                        for m in self.library_models:
-                            tags_response = client.get(_base_url.join(f"{m}/tags"))
-                            tags_response.raise_for_status()
-                            parsed_tags_html = lxml.html.document_fromstring(
-                                tags_response.text
-                            )
-                            model_unique_tags: Set[str] = set()
-                            for _, attribute, link, _ in lxml.html.iterlinks(
-                                parsed_tags_html
-                            ):
-                                if attribute == "href" and link.startswith(
-                                    f"/library/{m}:"
-                                ):
-                                    if m not in self.models_tags:
-                                        self.models_tags[m] = []
-                                    model_unique_tags.add(link.replace("/library/", ""))
-                            self.models_tags[m] = list(model_unique_tags)
-                            model_counter += 1
-                            progress.update(tags_task, completed=model_counter)
-                    logger.info(
-                        f"Updated {len(self.models_tags)} models with tags from the Ollama library."
+            self.settings.ollama_library.verify_ssl,
+            self.settings.ollama_library.timeout,
+        ) as http_client:
+            response = http_client.get(url)
+            response.raise_for_status()
+            return response.text
+
+    def _make_blob_url(
+        self, model_identifier: str, digest: str, model_source: ModelSource
+    ) -> httpx.URL:
+        """
+        Constructs the blob URL for a given model and digest.
+
+        Args:
+            model_identifier (str): The model name, e.g., "gpt-oss:latest".
+            digest (str): The digest of the blob, e.g., "sha256:abcdef...".
+            model_source (ModelSource): The source of the model (e.g., OLLAMA or HUGGINGFACE).
+
+        Returns:
+            httpx.URL: The constructed blob URL.
+        """
+        logger.debug(
+            f"Constructing blob URL for {model_identifier} with digest {digest}"
+        )
+        model_name = model_identifier.split(":")[0]
+        match model_source:
+            case ModelSource.OLLAMA:
+                return httpx.URL(self.settings.ollama_library.registry_base_url).join(
+                    f"{model_name}/blobs/{digest.replace(':', '-')}"
+                )
+            case ModelSource.HUGGINGFACE:
+                return httpx.URL(
+                    f"{ModelDownloader.HF_BASE_URL}{model_name}/blobs/{digest}"
+                )
+            case _:
+                raise ValueError(f"Unsupported model source: {model_source}")
+
+    def _download_model_blob(
+        self, model_identifier: str, named_digest: str, model_source: ModelSource
+    ) -> tuple:
+        """
+        Downloads a model blob given its named digest.
+
+        Args:
+            model_identifier (str): The model name, e.g., "gpt-oss:latest".
+            named_digest (str): The named digest of the blob, e.g., "sha256:abcdef...".
+            model_source (ModelSource): The source of the model (e.g., OLLAMA or HUGGINGFACE).
+
+        Returns:
+            tuple: A tuple containing the path to the downloaded file as string and its digest.
+        """
+        url = self._make_blob_url(
+            model_identifier=model_identifier,
+            digest=named_digest,
+            model_source=model_source,
+        )
+        # try:
+        sha256_hash = hashlib.new("sha256")
+        with tempfile.NamedTemporaryFile(delete=False) as temp_file:
+            self._unnecessary_files.add(temp_file.name)
+            with self.get_httpx_client(
+                self.settings.ollama_library.verify_ssl,
+                self.settings.ollama_library.timeout,
+            ).stream("GET", url) as response:
+                response.raise_for_status()
+                total = int(response.headers["Content-Length"])
+
+                with Progress(
+                    TextColumn(text_format="{task.description}"),
+                    "[progress.percentage]{task.percentage:>3.0f}%",
+                    BarColumn(bar_width=None),
+                    DownloadColumn(),
+                    TransferSpeedColumn(),
+                ) as progress:
+                    download_task = progress.add_task(
+                        f"Downloading BLOB {named_digest[:11]}...{named_digest[-4:]}",
+                        total=total,
                     )
-                    self._save_models_tags_cache()
-                return self.models_tags
+                    for chunk in response.iter_bytes():
+                        sha256_hash.update(chunk)
+                        temp_file.write(chunk)
+                        progress.update(
+                            download_task, completed=response.num_bytes_downloaded
+                        )
+        logger.debug(f"Downloaded {url} to {temp_file.name}")
+        content_digest = sha256_hash.hexdigest()
+        logger.debug(f"Computed SHA256 digest of {temp_file.name}: {content_digest}")
+        return temp_file.name, content_digest
+
+    def _save_manifest(
+        self, data: str, model_identifier: str, model_source: ModelSource
+    ):
+        """
+        Saves the manifest data to the appropriate location based on the model source.
+
+        Args:
+            data (str): The manifest JSON content as a string.
+            model_identifier (str): The model identifier, e.g., "gpt-oss:latest" for an Ollama library model
+            or "unsloth/gemma-3-270m-it-GGUF:Q4_K_M" for a Hugging Face model.
+            model_source (ModelSource): The source of the model (e.g., OLLAMA or HUGGINGFACE).
+
+        Returns:
+            str: The path to the saved manifest file.
+        """
+        manifests_toplevel_dir = os.path.join(
+            (
+                os.path.expanduser(self.settings.ollama_library.models_path)
+                if self.settings.ollama_library.models_path.startswith("~")
+                else self.settings.ollama_library.models_path
+            ),
+            "manifests",
+        )
+        model_identifier_splits = model_identifier.split(":")
+        manifests_dir = None
+        model_source_host = None
+        match model_source:
+            case ModelSource.OLLAMA:
+                model_source_host = urlparse(
+                    self.settings.ollama_library.registry_base_url
+                ).hostname
+                manifests_dir = os.path.join(
+                    manifests_toplevel_dir,
+                    model_source_host,
+                    "library",
+                    model_identifier_splits[0],
+                )
+            case ModelSource.HUGGINGFACE:
+                model_source_host = urlparse(ModelDownloader.HF_BASE_URL).hostname
+                manifests_dir = os.path.join(
+                    manifests_toplevel_dir,
+                    # FIXME: This is a hack for mypy -- the model_source_host should not be None
+                    model_source_host or "",
+                    model_identifier_splits[0],
+                )
+            case _:
+                raise ValueError(f"Unsupported model source: {model_source}")
+        if not os.path.exists(manifests_dir):
+            logger.warning(
+                f"Manifests path {manifests_dir} does not exist. Will attempt to create it."
+            )
+            os.makedirs(manifests_dir)
+            self._unnecessary_files.add(manifests_dir)
+        target_file = os.path.join(manifests_dir, model_identifier_splits[1])
+        with open(target_file, "w") as f:
+            f.write(data)
+            logger.info(f"Saved manifest to {target_file}")
+        if self.settings.ollama_library.user_group:
+            user, group = self.settings.ollama_library.user_group
+            shutil.chown(target_file, user, group)
+            # The directory ownership must also be changed because it may have been created by a different user, most likely a sudoer
+            # TODO: Is this necessary or can the ownership change to the top-level directory cascade down?
+            shutil.chown(manifests_dir, user, group)
+            shutil.chown(manifests_toplevel_dir, user, group)
+            logger.info(
+                f"Changed ownership of {target_file} to user: {user}, group: {group}"
+            )
+        self._unnecessary_files.add(target_file)
+        return target_file
+
+    def _save_blob(
+        self,
+        source: str,
+        named_digest: str,
+        computed_digest: str,
+    ) -> Tuple[bool, str | None]:
+        """
+        Saves the downloaded blob to the appropriate location based on the model source.
+
+        Args:
+            source (str): The path to the downloaded BLOB.
+            named_digest (str): The expected digest of the BLOB prefixed with the digest algorithm followed by the colon character.
+            computed_digest (str): The computed digest of the BLOB.
+
+        Returns:
+            Tuple[bool, str | None]: A tuple containing a boolean indicating success or failure,
+            and the path to the saved BLOB if successful, None otherwise.
+        """
+        if computed_digest != named_digest[7:]:
+            logger.error(
+                f"Digest mismatch: expected {named_digest[7:]}, got {computed_digest}"
+            )
+            return False, None
+
+        blobs_dir = os.path.join(
+            (
+                os.path.expanduser(self.settings.ollama_library.models_path)
+                if self.settings.ollama_library.models_path.startswith("~")
+                else self.settings.ollama_library.models_path
+            ),
+            "blobs",
+        )
+
+        logger.info(f"BLOB {named_digest} digest verified successfully.")
+        if not os.path.isdir(blobs_dir):
+            logger.error(f"BLOBS path {blobs_dir} must be a directory.")
+            return False, None
+        if not os.path.exists(blobs_dir):
+            logger.error(f"BLOBS path {blobs_dir} must exist.")
+            return False, None
+        target_file = os.path.join(blobs_dir, named_digest.replace(":", "-"))
+        shutil.move(source, target_file)
+        logger.info(f"Moved {source} to {target_file}")
+        if self.settings.ollama_library.user_group:
+            user, group = self.settings.ollama_library.user_group
+            shutil.chown(target_file, user, group)
+            shutil.chown(blobs_dir, user, group)
+            # Set permissions to rw-r-----
+            os.chmod(target_file, 0o640)
+            logger.info(
+                f"Changed ownership of {target_file} to user: {user}, group: {group}"
+            )
+        self._unnecessary_files.remove(source)
+        self._unnecessary_files.add(target_file)
+        return True, target_file
+
+    def cleanup_unnecessary_files(self):
+        """
+        Cleans up unnecessary files and directories created during downloading models.
+        """
+        # TODO: Is this thread-safe? Should we use a lock?
+        if not self._cleanup_running:
+            self._cleanup_running = True
+            list_of_unnecessary_files = list(self._unnecessary_files)
+            unnecessary_directories = set()
+            for file_object in list_of_unnecessary_files:
+                try:
+                    if not os.path.isdir(file_object):
+                        os.remove(file_object)
+                        logger.info(f"Removed unnecessary file: {file_object}")
+                    else:
+                        # If it's a directory, we don't remove it yet because it may not be empty.
+                        unnecessary_directories.add(file_object)
+                    self._unnecessary_files.remove(file_object)
+                except Exception as e:
+                    logger.error(
+                        f"Failed to remove unnecessary file {file_object}: {e}"
+                    )
+
+            # Now remove unnecessary directories if they are empty
+            for directory in unnecessary_directories:
+                try:
+                    os.rmdir(directory)
+                    logger.info(f"Removed unnecessary directory: {directory}")
+                except OSError as e:
+                    logger.error(
+                        f"Failed to remove unnecessary directory {directory}: {e}"
+                    )
+            self._cleanup_running = False
